+++ etl/transform_positives.py
@@ -6,10 +6,11 @@ storage folder "positives + augmentation."
 
 import logging
 # from matplotlib import pyplot as plt
-from lib import transforms, cloud_management as cloud
+from lib import cloud_management as cloud
 import numpy as np
 import pandas as pd
 import itertools
+
 
 def configure_logger():
     root_logger = logging.getLogger()
@@ -20,8 +21,9 @@ def configure_logger():
     handler.setFormatter(formatter)
     root_logger.addHandler(handler)
 
+
 def transform_one(arr, file_id):
-    iterlist = list(itertools.product('01', repeat = 3))
+    iterlist = list(itertools.product('01', repeat=3))
     axes = [[0, 1], [1, 2], [0, 2]]
 
     transform_number = 0
@@ -39,7 +41,9 @@ def transform_one(arr, file_id):
                 # save to the numpy generator source directory
             file_id_new =  file_id + "_" + str(transform_number)
             # print(file_id_new)
-            cloud.save_chunks_to_cloud(flipped, 'normal/positive', file_id_new)
+            cloud.save_chunks_to_cloud(
+                flipped, 'normal/positive', file_id_new)
+
 
 def transform_positives():
     configure_logger()
@@ -63,13 +67,14 @@ def transform_positives():
         input_arr = cloud.download_array(in_blob)
         logging.info(f"blob shape: {input_arr.shape}")
         # crop individual input array
-        flipped = transform_one(input_arr, file_id)
+        transform_one(input_arr, file_id)
 
         # plt.figure(figsize=(6, 6))
         # plt.imshow(not_extreme_arr, interpolation='none')
         # plt.show()
 
-def clean_data():
+
+def clean_old_data():
     configure_logger()
     client = cloud.authenticate()
     bucket = client.get_bucket('elvos')
@@ -81,19 +86,45 @@ def clean_data():
     i = 0
     for in_blob in bucket.list_blobs(prefix=prefix):
         i += 1
-        # # blacklist
-        # if in_blob.name == prefix + 'LAUIHISOEZIM5ILF.npy':
-        #     continue
-        #
-        # # perform the normal cropping procedure
-        # logging.info(f'downloading {in_blob.name}')
-        # file_id = in_blob.name.split('/')[-1]
-        # file_id = file_id.split('.')[0]
-        #
-        # if '_' in file_id:
-        #     continue
-        #
-        # in_blob.delete()
+        # blacklist
+        if in_blob.name == prefix + 'LAUIHISOEZIM5ILF.npy':
+            continue
+
+        # perform the normal cropping procedure
+        logging.info(f'downloading {in_blob.name}')
+        file_id = in_blob.name.split('/')[-1]
+        file_id = file_id.split('.')[0]
+
+        if '_' in file_id:
+            in_blob.delete()
+    print(i)
+
+
+def clean_new_data():
+    configure_logger()
+    client = cloud.authenticate()
+    bucket = client.get_bucket('elvos')
+
+    # iterate through every source directory...
+    prefix = "chunk_data/normal/positive"
+    logging.info(f"transforming positive chunks from {prefix}")
+
+    i = 0
+    for in_blob in bucket.list_blobs(prefix=prefix):
+        i += 1
+        # blacklist
+        if in_blob.name == prefix + 'LAUIHISOEZIM5ILF.npy':
+            continue
+
+        # perform the normal cropping procedure
+        logging.info(f'downloading {in_blob.name}')
+        file_id = in_blob.name.split('/')[-1]
+        file_id = file_id.split('.')[0]
+
+        if '_' in file_id:
+            continue
+
+        in_blob.delete()
     print(i)
 
 
@@ -103,14 +134,16 @@ def generate_csv():
     labels_df = pd.read_csv('/home/amy/data/annotated_labels.csv')
     for index, row in labels_df.iterrows():
         print(index, row[1])
-        if(row[1]== 1):
+        if(row[1] == 1):
             # every time you come across a positive, add in 24 more rows
             to_add = {}
             for i in range(24):
                 # we use index + 1 because filenames are 1-indexed
                 new_patient_id = str(row[0]) + "_" + str(i + 1)
                 to_add[index + 500000 + i] = [new_patient_id, 1]
-            to_add_df = pd.DataFrame.from_dict(to_add, orient='index', columns =['Unnamed: 0', 'label'])
+            to_add_df = pd.DataFrame.from_dict(to_add, orient='index',
+                                               columns=[
+                                                   'Unnamed: 0', 'label'])
             print(to_add_df)
             labels_df = labels_df.append(to_add_df)
     # print(labels_df.loc['04IOS24JP70LHBGB184', 'Unnamed: 0'])
@@ -119,8 +152,13 @@ def generate_csv():
     labels_df.to_csv("augmented_annotated_labels.csv")
 
 
-if __name__ == '__main__':
+def run_transform():
+    clean_old_data()
     configure_logger()
-    # generate_csv()
-    # transform_positives()
-    clean_data()
+    generate_csv()
+    transform_positives()
+    clean_new_data()
+
+
+if __name__ == '__main__':
+    run_transform()
